more swerve math
